stuff migrated from engine to adapter
