+++ IFC_Adapter/Methods/Convert/FromIFC.cs
@@ -22,22 +22,49 @@
 
 using BH.oM.Adapters.IFC;
 using BH.oM.Base;
-using BH.oM.Physical.Elements;
 using System.Collections.Generic;
 using Xbim.Ifc2x3.Interfaces;
 
-namespace BH.Engine.Adapters.IFC
+namespace BH.Adapter.IFC
 {
     public static partial class Convert
     {
         /***************************************************/
+        /****             Interface Methods             ****/
+        /***************************************************/
+
+        public static IEnumerable<IBHoMObject> IFromIFC(this IIfcElement element, Discipline discipline, IFCSettings settings = null)
+        {
+            IEnumerable<IBHoMObject> result = FromIFC(element as dynamic, discipline, settings);
+            if (result == null)
+                BH.Engine.Reflection.Compute.RecordError($"IFC element conversion to BHoM failed for discipline {discipline}.");
+
+            return result;
+        }
+
+
+        /***************************************************/
         /****              Public Methods               ****/
         /***************************************************/
 
-        public static Floor FloorFromIFC(this IIfcSlab element, IFCSettings settings)
+        public static IEnumerable<IBHoMObject> FromIFC(this IIfcSlab element, Discipline discipline, IFCSettings settings)
         {
-            //TODO: refine this!
-            return new Floor();
+            switch (discipline)
+            {
+                default:
+                    return new List<IBHoMObject> { element.FloorFromIFC(settings) };
+            }
+        }
+
+        /***************************************************/
+
+        public static IEnumerable<IBHoMObject> FromIFC(this IIfcReinforcingBar element, Discipline discipline, IFCSettings settings)
+        {
+            switch (discipline)
+            {
+                default:
+                    return new List<IBHoMObject> { element.ReinforcingBarFromIFC(settings) };
+            }
         }
 
         /***************************************************/
