+++ IFC_Adapter/Methods/Convert/Geometry/FromIFC/TransformMatrix.cs
@@ -23,7 +23,7 @@
 using BH.oM.Geometry;
 using Xbim.Common.Geometry;
 
-namespace BH.Engine.Adapters.IFC
+namespace BH.Adapter.IFC
 {
     public static partial class Convert
     {
@@ -31,9 +31,29 @@ namespace BH.Engine.Adapters.IFC
         /****              Public Methods               ****/
         /***************************************************/
 
-        public static Point PointFromIFC(this XbimPoint3D point)
+        public static TransformMatrix TransformMatrixFromIFC(this XbimMatrix3D matrix)
         {
-            return new Point { X = point.X, Y = point.Y, Z = point.Z };
+            double[] dbls = matrix.ToDoubleArray();
+            TransformMatrix result = new TransformMatrix();
+
+            result.Matrix[0, 0] = dbls[0];
+            result.Matrix[1, 0] = dbls[1];
+            result.Matrix[2, 0] = dbls[2];
+            result.Matrix[3, 0] = dbls[3];
+            result.Matrix[0, 1] = dbls[4];
+            result.Matrix[1, 1] = dbls[5];
+            result.Matrix[2, 1] = dbls[6];
+            result.Matrix[3, 1] = dbls[7];
+            result.Matrix[0, 2] = dbls[8];
+            result.Matrix[1, 2] = dbls[9];
+            result.Matrix[2, 2] = dbls[10];
+            result.Matrix[3, 2] = dbls[11];
+            result.Matrix[0, 3] = dbls[12];
+            result.Matrix[1, 3] = dbls[13];
+            result.Matrix[2, 3] = dbls[14];
+            result.Matrix[3, 3] = dbls[15];
+
+            return result;
         }
 
         /***************************************************/
