+++ IFC_Adapter/Methods/Convert/Physical/FromIFC/Floor.cs
@@ -20,10 +20,13 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using BH.oM.Geometry;
-using Xbim.Common.Geometry;
+using BH.oM.Adapters.IFC;
+using BH.oM.Base;
+using BH.oM.Physical.Elements;
+using System.Collections.Generic;
+using Xbim.Ifc2x3.Interfaces;
 
-namespace BH.Engine.Adapters.IFC
+namespace BH.Adapter.IFC
 {
     public static partial class Convert
     {
@@ -31,29 +34,10 @@ namespace BH.Engine.Adapters.IFC
         /****              Public Methods               ****/
         /***************************************************/
 
-        public static TransformMatrix TransformMatrixFromIFC(this XbimMatrix3D matrix)
+        public static Floor FloorFromIFC(this IIfcSlab element, IFCSettings settings)
         {
-            double[] dbls = matrix.ToDoubleArray();
-            TransformMatrix result = new TransformMatrix();
-
-            result.Matrix[0, 0] = dbls[0];
-            result.Matrix[1, 0] = dbls[1];
-            result.Matrix[2, 0] = dbls[2];
-            result.Matrix[3, 0] = dbls[3];
-            result.Matrix[0, 1] = dbls[4];
-            result.Matrix[1, 1] = dbls[5];
-            result.Matrix[2, 1] = dbls[6];
-            result.Matrix[3, 1] = dbls[7];
-            result.Matrix[0, 2] = dbls[8];
-            result.Matrix[1, 2] = dbls[9];
-            result.Matrix[2, 2] = dbls[10];
-            result.Matrix[3, 2] = dbls[11];
-            result.Matrix[0, 3] = dbls[12];
-            result.Matrix[1, 3] = dbls[13];
-            result.Matrix[2, 3] = dbls[14];
-            result.Matrix[3, 3] = dbls[15];
-
-            return result;
+            //TODO: refine this!
+            return new Floor();
         }
 
         /***************************************************/
