+++ IFC_Engine/IFC_Engine.csproj
@@ -201,18 +201,10 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Convert\FromIFC.cs" />
-    <Compile Include="Convert\Geometry\FromIFC\Point.cs" />
-    <Compile Include="Convert\Geometry\FromIFC\TransformMatrix.cs" />
-    <Compile Include="Convert\Physical\FromIFC\Floor.cs" />
-    <Compile Include="Convert\Physical\FromIFC\ReinforcingBar.cs" />
-    <Compile Include="Query\CorrespondentIFCTypes.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Query\DefaultIfNull.cs" />
     <Compile Include="Query\Discipline.cs" />
-    <Compile Include="Query\IFCEntities.cs" />
     <Compile Include="Query\IFCMeshes.cs" />
-    <Compile Include="Query\Meshes.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Compute\" />
